+++ lib/middleman-akcms/archive.rb
@@ -18,7 +18,7 @@ module Middleman::Akcms
 
       @controller.articles.group_by {|a| 
         Date.new(a.date.year, a.date.month, 1)}.each {|date_ym, articles|
-        @archives[date_ym] = create_proxy_resource(:date, date_ym, articles: articles)
+        @archives[date_ym] = create_proxy_resource(link(date_ym), date: date_ym, articles: articles)
       }
       return resources + @archives.values.sort_by {|res| res.locals[:date]}.reverse
     end